webstorage/storage-clear-empty-storage.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebStorage Test: Storage - clear() empty storage</title>
<link rel="author" title="Intel" href="http://www.intel.com" >
<link rel="help" href="https://w3c.github.io/webstorage/#the-storage-interface" >
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  ["localStorage", "sessionStorage"].forEach(function(name) {
    test(function () {
      var storage = window[name];
      storage.clear();

      storage.setItem(null, "value");
      storage.removeItem(null);
      assert_equals(storage.length, 0, "The element storage.length should be 0");

      storage.clear();
      assert_equals(storage.length, 0, "storage.length should be 0");

    }, name + ": When no key/value pairs, storage.clear method does nothing");
  });
</script>
